requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-groq
langchain-mistralai
langgraph
python-dotenv
faiss-cpu
tiktoken
fastapi
uvicorn
requests
langchain_huggingface


# The below libraries if local_model.py doesnt run
transformers
torch
torchvision
torchaudio

sentence-transformers